BOT-31: многа апдейтов

+++ internal/delivery_http/posts/posts_query/get_by_id.go
@@ -11,7 +11,7 @@ import (
 func (p *PostsQueryHandler) GetPostById(ctx context.Context, params gen.GetPostByIdParams) (gen.GetPostByIdRes, error) {
 	// TODO: По идее тут должна быть бизнес логика выбора какой пост фетчить (с категориями или без) в зависимости от платформы,
 	// Пока возвращается пост без категорий
-	post, err := p.repo.GetByID(ctx, uint64(params.PostId))
+	post, err := p.repo.GetByID(ctx, params.PostId)
 	if err != nil {
 		return &gen.GetPostByIdInternalServerError{ErrorCode: http.StatusInternalServerError, Message: err.Error()}, nil
 	}
